refactoring the inner functions out of the peakfinders
