+++ pyxem/utils/peakfinders2D.py
@@ -18,6 +18,9 @@
 
 import numpy as np
 import scipy.ndimage as ndi
+from pyxem.utils.peakfinder_utils import *
+
+
 
 NO_PEAKS = np.array([[[np.nan, np.nan]]])
 
@@ -59,45 +62,6 @@ def find_peaks_zaefferer(z, grad_threshold=0.1, window_size=40,
     crystallographic analysis in transmission electron microscopy" J. Ap. Cryst.
     This version by Ben Martineau (2016)
     """
-
-    def box(x, y, window_size, x_max, y_max):
-        """Produces a list of coordinates in the box about (x, y)."""
-        a = int(window_size / 2)
-        x_min = max(0, x - a)
-        x_max = min(x_max, x + a)
-        y_min = max(0, y - a)
-        y_max = min(y_max, y + a)
-        return np.array(
-            np.meshgrid(range(x_min, x_max), range(y_min, y_max))).reshape(
-            2, -1).T
-
-    def get_max(image, box):
-        """Finds the coordinates of the maximum of 'image' in 'box'."""
-        vals = image[box[:, 0], box[:, 1]]
-        max_position = box[np.argmax(vals)]
-        return max_position
-
-    def distance(x, y):
-        """Calculates the distance between two points."""
-        v = x - y
-        return np.sqrt(np.sum(np.square(v)))
-
-    def gradient(image):
-        """Calculates the square of the 2-d partial gradient.
-
-        Parameters
-        ----------
-        image : numpy.ndarray
-
-        Returns
-        -------
-        numpy.ndarray
-
-        """
-        gradient_of_image = np.gradient(image)
-        gradient_of_image = gradient_of_image[0] ** 2 + gradient_of_image[
-                                                            1] ** 2
-        return gradient_of_image
 
     # Generate an ordered list of matrix coordinates.
     if len(z.shape) != 2:
@@ -157,92 +121,21 @@ def find_peaks_stat(z, alpha=1., window_radius=10, convergence_ratio=0.05):
     Cambridge.
     This version by Ben Martineau (2016), with minor modifications to the
     original where methods were ambiguous or unclear.
+
+    Algorithm stags in the comments
     """
-    from scipy.ndimage.filters import generic_filter
-    from scipy.ndimage.filters import uniform_filter
-    from sklearn.cluster import DBSCAN
+    image = normalize(z)  # 1
+    image = stat_binarise(image,window_radius,alpha)  # 2, 3
+    n_peaks = np.infty  # Initial number of peaks
+    image, peaks = peak_find_once(image)  # 4-6
+    m_peaks = len(peaks)  # Actual number of peaks
+    while (n_peaks - m_peaks) / n_peaks > convergence_ratio:  # 8
+        n_peaks = m_peaks
+        image, peaks = peak_find_once(image)
+        m_peaks = len(peaks)
+    peak_centers = np.array([np.mean(peak, axis=0) for peak in peaks])  # 7
 
-    def normalize(image):
-        """Scales the image to intensities between 0 and 1."""
-        return image / np.max(image)
-
-    def _local_stat(image, radius, func):
-        """Calculates rolling method 'func' over a circular kernel."""
-        x, y = np.ogrid[-radius:radius + 1, -radius:radius + 1]
-        kernel = np.hypot(x, y) < radius
-        stat = generic_filter(image, func, footprint=kernel)
-        return stat
-
-    def local_mean(image, radius):
-        """Calculates rolling mean over a circular kernel."""
-        return _local_stat(image, radius, np.mean)
-
-    def local_std(image, radius):
-        """Calculates rolling standard deviation over a circular kernel."""
-        return _local_stat(image, radius, np.std)
-
-    def single_pixel_desensitize(image):
-        """Reduces single-pixel anomalies by nearest-neighbor smoothing."""
-        kernel = np.array([[0.5, 1, 0.5], [1, 1, 1], [0.5, 1, 0.5]])
-        smoothed_image = generic_filter(image, np.mean, footprint=kernel)
-        return smoothed_image
-
-    def stat_binarise(image):
-        """Peaks more than one standard deviation from the mean set to one."""
-        image_rolling_mean = local_mean(image, window_radius)
-        image_rolling_std = local_std(image, window_radius)
-        image = single_pixel_desensitize(image)
-        binarised_image = np.zeros(image.shape)
-        stat_mask = image > (image_rolling_mean + alpha * image_rolling_std)
-        binarised_image[stat_mask] = 1
-        return binarised_image
-
-    def smooth(image):
-        """Image convolved twice using a uniform 3x3 kernel."""
-        image = uniform_filter(image, size=3)
-        image = uniform_filter(image, size=3)
-        return image
-
-    def half_binarise(image):
-        """Image binarised about values of one-half intensity."""
-        binarised_image = np.where(image > 0.5, 1, 0)
-        return binarised_image
-
-    def separate_peaks(binarised_image):
-        """Identify adjacent 'on' coordinates via DBSCAN."""
-        bi = binarised_image.astype('bool')
-        coordinates = np.indices(bi.shape).reshape(2, -1).T[
-            bi.flatten()]
-        db = DBSCAN(2, 3)
-        peaks = []
-        labeled_points = db.fit_predict(coordinates)
-        for peak_label in list(set(labeled_points)):
-            peaks.append(coordinates[labeled_points == peak_label])
-        return peaks
-
-    def _peak_find_once(image):
-        """Smooth, binarise, and find peaks according to main algorithm."""
-        image = smooth(image)
-        image = half_binarise(image)
-        peaks = separate_peaks(image)
-        return image, peaks
-
-    def stat_peak_finder(image):
-        """Find peaks in image. Algorithm stages in comments."""
-        image = normalize(image)  # 1
-        image = stat_binarise(image)  # 2, 3
-        n_peaks = np.infty  # Initial number of peaks
-        image, peaks = _peak_find_once(image)  # 4-6
-        m_peaks = len(peaks)  # Actual number of peaks
-        while (n_peaks - m_peaks) / n_peaks > convergence_ratio:  # 8
-            n_peaks = m_peaks
-            image, peaks = _peak_find_once(image)
-            m_peaks = len(peaks)
-        peak_centers = np.array(
-            [np.mean(peak, axis=0) for peak in peaks])  # 7
-        return peak_centers
-
-    return clean_peaks(stat_peak_finder(z))
+    return clean_peaks(peak_centers)
 
 
 def find_peaks_dog(z, min_sigma=1., max_sigma=50., sigma_ratio=1.6,
